requirements.txt
numpy
pandas
python-dotenv
pymysql
scikit-learn
seaborn
catboost
xgboost

#-e .